add ( new_password ) : add modalNewPassword
