+++ lib/presentation/screens/login_v2/new_password_v2.dart
@@ -1,15 +1,12 @@
-import 'package:finniu/presentation/providers/recovery_password_provider.dart';
-import 'package:finniu/presentation/providers/settings_provider.dart';
-import 'package:finniu/presentation/providers/user_provider.dart';
 import 'package:finniu/presentation/screens/catalog/widgets/inputs_user_v2/input_password_v2.dart';
 import 'package:finniu/presentation/screens/catalog/widgets/send_proof_button.dart';
 import 'package:finniu/presentation/screens/catalog/widgets/snackbar/snackbar_v2.dart';
 import 'package:finniu/presentation/screens/catalog/widgets/text_poppins.dart';
 import 'package:finniu/presentation/screens/catalog/widgets/user_profil_v2/scafold_user_profile.dart';
+import 'package:finniu/presentation/screens/login_v2/helpers/new_password_helper.dart';
+import 'package:finniu/presentation/screens/login_v2/widgets/password_new_requiered.dart';
 import 'package:finniu/presentation/screens/v2_user_profile/helpers/validate_form.dart';
-import 'package:finniu/presentation/screens/v2_user_profile/widgets/password_required.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:loader_overlay/loader_overlay.dart';
 
@@ -87,35 +84,11 @@ class FormNewPasword extends HookConsumerWidget {
         if (passwordError.value) return;
         if (passwordConfirmError.value) return;
         context.loaderOverlay.show();
-        final status = await ref.read(
-          setNewPasswordFutureProvider(passwordController.text).future,
+        resetPassword(
+          context: context,
+          ref: ref,
+          newPassword: passwordController.text.trim(),
         );
-        if (status == true) {
-          ref
-              .read(userProfileNotifierProvider.notifier)
-              .setPassword(passwordController.text);
-          const SnackBar(
-            content: Text('Contraseña cambiada con éxito'),
-          );
-          showSnackBarV2(
-            context: context,
-            title: "Contraseña cambiada con ≠!",
-            message: 'Contraseña cambiada con éxito!',
-            snackType: SnackType.success,
-          );
-
-          await Future.delayed(const Duration(seconds: 1));
-          Navigator.pushNamed(context, '/v2/login_email');
-          context.loaderOverlay.hide();
-        } else {
-          context.loaderOverlay.hide();
-          showSnackBarV2(
-            context: context,
-            title: "Error al cambiar la contraseña",
-            message: 'No se pudo cambiar la contraseña',
-            snackType: SnackType.error,
-          );
-        }
       }
     }
 
@@ -201,269 +174,3 @@ class FormNewPasword extends HookConsumerWidget {
     );
   }
 }
-
-class PasswordNewRequired extends HookConsumerWidget {
-  const PasswordNewRequired({
-    super.key,
-    required this.controller,
-    required this.isExpanded,
-  });
-  final bool isExpanded;
-  final TextEditingController controller;
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final isDarkMode = ref.watch(settingsNotifierProvider).isDarkMode;
-    const int containerDark = 0xff222222;
-    const int containerLight = 0xffF7F7F7;
-
-    final ValueNotifier<bool> eightCharacters = ValueNotifier<bool>(false);
-    final ValueNotifier<bool> oneNumber = ValueNotifier<bool>(false);
-    final ValueNotifier<bool> oneLowercase = ValueNotifier<bool>(false);
-    final ValueNotifier<bool> oneCapital = ValueNotifier<bool>(false);
-    final ValueNotifier<bool> oneSpecialSymbol = ValueNotifier<bool>(false);
-
-    final RegExp hasDigits = RegExp(r'[0-9]');
-    final RegExp hasLowercase = RegExp(r'[a-z]');
-    final RegExp hasUppercase = RegExp(r'[A-Z]');
-    final RegExp hasSpecialCharacters = RegExp(r'[!@#$%^&*(),.?":{}|<>]');
-
-    useEffect(
-      () {
-        controller.text.length >= 8
-            ? eightCharacters.value = true
-            : eightCharacters.value = false;
-        hasDigits.hasMatch(controller.text)
-            ? oneNumber.value = true
-            : oneNumber.value = false;
-        hasLowercase.hasMatch(controller.text)
-            ? oneLowercase.value = true
-            : oneLowercase.value = false;
-        hasUppercase.hasMatch(controller.text)
-            ? oneCapital.value = true
-            : oneCapital.value = false;
-        hasSpecialCharacters.hasMatch(controller.text)
-            ? oneSpecialSymbol.value = true
-            : oneSpecialSymbol.value = false;
-        controller.addListener(() {
-          controller.text.length >= 8
-              ? eightCharacters.value = true
-              : eightCharacters.value = false;
-          hasDigits.hasMatch(controller.text)
-              ? oneNumber.value = true
-              : oneNumber.value = false;
-          hasLowercase.hasMatch(controller.text)
-              ? oneLowercase.value = true
-              : oneLowercase.value = false;
-          hasUppercase.hasMatch(controller.text)
-              ? oneCapital.value = true
-              : oneCapital.value = false;
-          hasSpecialCharacters.hasMatch(controller.text)
-              ? oneSpecialSymbol.value = true
-              : oneSpecialSymbol.value = false;
-        });
-        return null;
-      },
-      [controller],
-    );
-
-    return Positioned(
-      top: 120,
-      child: Container(
-        decoration: BoxDecoration(
-          borderRadius: BorderRadius.circular(10),
-          color: isDarkMode
-              ? const Color(containerDark)
-              : const Color(containerLight),
-        ),
-        width: MediaQuery.of(context).size.width * 0.9,
-        height: 200,
-        padding: const EdgeInsets.all(10),
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.spaceAround,
-          children: [
-            ValueListenableBuilder<bool>(
-              valueListenable: eightCharacters,
-              builder: (context, value, child) {
-                return ItemValidate(
-                  text: "Al menos 8 caracteres",
-                  isValidate: eightCharacters.value,
-                );
-              },
-            ),
-            ValueListenableBuilder<bool>(
-              valueListenable: oneNumber,
-              builder: (context, value, child) {
-                return ItemValidate(
-                  text: "Al menos 1 número (0....9)",
-                  isValidate: oneNumber.value,
-                );
-              },
-            ),
-            ValueListenableBuilder<bool>(
-              valueListenable: oneLowercase,
-              builder: (context, value, child) {
-                return ItemValidate(
-                  text: "Al menos 1 letra minúscula (a...z)",
-                  isValidate: oneLowercase.value,
-                );
-              },
-            ),
-            ValueListenableBuilder<bool>(
-              valueListenable: oneCapital,
-              builder: (context, value, child) {
-                return ItemValidate(
-                  text: "Al menos 1 letra mayúscula (A...Z)",
-                  isValidate: oneCapital.value,
-                );
-              },
-            ),
-            ValueListenableBuilder<bool>(
-              valueListenable: oneSpecialSymbol,
-              builder: (context, value, child) {
-                return ItemValidate(
-                  text: "Al menos 1 símbolo especial (!...\$)",
-                  isValidate: oneSpecialSymbol.value,
-                );
-              },
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-
-// class NewPasswordV2 extends HookConsumerWidget {
-//   const NewPasswordV2({super.key});
-
-//   @override
-//   Widget build(BuildContext context, WidgetRef ref) {
-//     final themeProvider = ref.watch(settingsNotifierProvider);
-//     final passwordController = useTextEditingController();
-//     // final currentTheme = Provider.of<SettingsProvider>(context, listen: false);
-//     final isHidden = useState(true);
-//     return CustomScaffoldReturn(
-//       body: SingleChildScrollView(
-//         child: Center(
-//           child: Column(
-//             mainAxisAlignment: MainAxisAlignment.center,
-//             children: <Widget>[
-//               SizedBox(height: MediaQuery.of(context).size.height * 0.1),
-//               Container(
-//                 alignment: Alignment.topCenter,
-//                 child: SizedBox(
-//                   child: Image.asset(
-//                     'assets/forgotpassword/padlock.png',
-//                     width: 67,
-//                     height: 75,
-//                   ),
-//                 ),
-//               ),
-//               TextPoppins(
-//                 text: 'Cambiar tu contraseña',
-//                 colorText: Theme.of(context).textTheme.titleLarge!.color!.value,
-//                 fontSize: 24,
-//                 fontWeight: FontWeight.w600,
-//               ),
-//               const SizedBox(height: 20),
-//               TextPoppins(
-//                 text: 'Ingresa tu nueva contraseña',
-//                 colorText: themeProvider.isDarkMode ? (whiteText) : (blackText),
-//                 fontSize: 12,
-//                 fontWeight: FontWeight.w500,
-//               ),
-//               const SizedBox(height: 20),
-//               SizedBox(
-//                 width: 224,
-//                 height: 38,
-//                 child: TextFormField(
-//                   // onChanged: (value) {
-//                   //   _password = value;
-//                   // },
-//                   validator: (value) {
-//                     if (value == null || value.isEmpty) {
-//                       return 'Por favor ingrese algún valor';
-//                     }
-//                     return null;
-//                   },
-//                   obscureText: isHidden.value, // esto oculta la contrasenia
-//                   obscuringCharacter:
-//                       '*', //el caracter el cual reemplaza la contrasenia
-//                   decoration: InputDecoration(
-//                     suffixIconConstraints: const BoxConstraints(
-//                       maxHeight: 38,
-//                       minWidth: 38,
-//                     ),
-//                     suffixIcon: IconButton(
-//                       splashRadius: 20,
-//                       padding: EdgeInsets.zero,
-//                       icon: Icon(
-//                         isHidden.value
-//                             ? Icons.visibility
-//                             : Icons.visibility_off,
-//                         size: 23.20,
-//                       ),
-//                       alignment: Alignment.center,
-//                       onPressed: () {
-//                         isHidden.value = !isHidden.value;
-//                       },
-//                     ),
-//                     label: const Text(
-//                       "Contraseña",
-//                     ),
-//                   ),
-//                   controller: passwordController,
-//                 ),
-//               ),
-//               const SizedBox(height: 25),
-//               Container(
-//                 height: 50,
-//                 width: 224,
-//                 decoration: BoxDecoration(
-//                   borderRadius: BorderRadius.circular(50),
-//                   color: const Color(primaryDark),
-//                 ),
-//                 child: Center(
-//                   child: TextButton(
-//                     onPressed: () async {
-//                       final status = await ref.read(
-//                         setNewPasswordFutureProvider(passwordController.text)
-//                             .future,
-//                       );
-//                       if (status == true) {
-//                         ref
-//                             .read(userProfileNotifierProvider.notifier)
-//                             .setPassword(passwordController.text);
-//                         const SnackBar(
-//                           content: Text('Contraseña cambiada con éxito'),
-//                         );
-//                         showSnackBarV2(
-//                           context: context,
-//                           title: "Contraseña cambiada con ≠!",
-//                           message: 'Contraseña cambiada con éxito!',
-//                           snackType: SnackType.success,
-//                         );
-
-//                         await Future.delayed(const Duration(seconds: 1));
-//                         Navigator.pushNamed(context, '/v2/login_email');
-//                       } else {
-//                         showSnackBarV2(
-//                           context: context,
-//                           title: "Error al cambiar la contraseña",
-//                           message: 'No se pudo cambiar la contraseña',
-//                           snackType: SnackType.error,
-//                         );
-//                       }
-//                     },
-//                     child: const Text('Cambiar contraseña'),
-//                   ),
-//                 ),
-//               ),
-//             ],
-//           ),
-//         ),
-//       ),
-//     );
-//   }
-// }
